+++ src/ndv/views/_vispy/_vispy.py
@@ -15,7 +15,8 @@ from vispy.color import Color
 from vispy.util.quaternion import Quaternion
 
 from ndv._types import CursorType
-from ndv.views.bases import ArrayCanvas, filter_mouse_events
+from ndv.views._app import filter_mouse_events
+from ndv.views.bases import ArrayCanvas
 from ndv.views.bases.graphics._canvas_elements import (
     CanvasElement,
     ImageHandle,
