+++ src/ndv/views/bases/__init__.py
@@ -3,7 +3,7 @@ from ._lut_view import LutView
 from ._view_base import Viewable
 from .graphics._canvas import ArrayCanvas, HistogramCanvas
 from .graphics._canvas_elements import CanvasElement, ImageHandle, RoiHandle
-from .graphics._mouseable import Mouseable, filter_mouse_events
+from .graphics._mouseable import Mouseable
 
 __all__ = [
     "ArrayCanvas",
@@ -15,5 +15,4 @@ __all__ = [
     "Mouseable",
     "RoiHandle",
     "Viewable",
-    "filter_mouse_events",
 ]
